DEPLOY_SITE added calculation of 1v1 elo (also changed elo from single to double) + calculated 1v1 elos from battle history
svn path=/trunk/other/; revision=8535

+++ Zero-K.info/ContentService.asmx.cs
@@ -288,9 +288,9 @@ namespace ZeroKWeb
                 IsLobbyAdmin = acc.IsLobbyAdministrator,
                 IsZeroKAdmin= acc.IsZeroKAdmin,
                 Avatar = acc.Avatar,
-                Elo =acc.Elo,
+                Elo =(float)acc.Elo,
                 EffectiveElo = acc.EffectiveElo,
-                EloWeight = acc.EloWeight,
+                EloWeight = (float)acc.EloWeight,
                 FactionID = acc.FactionID??0,
                 FactionName = acc.Faction != null? acc.Faction.Name:null,
                 SpringieLevel = acc.SpringieLevel
